interpret: move discriminant reading and writing to separate file
